Break out CrashReportPersistence
